+++ src/util/sockutils.c
@@ -123,7 +123,7 @@ ssize_t knx_dgramsock_peek_knx(int sock) {
 	if (recvfrom(sock, buffer, KNX_HEADER_SIZE, MSG_PEEK,
 	    (struct sockaddr*) &sender, &sender_len) == KNX_HEADER_SIZE) {
 
-		uint16_t length;
+		size_t length;
 		if (knx_unpack_header(buffer, NULL, &length))
 			return length >= KNX_HEADER_SIZE ? length : 0;
 	}
